UltimateAV/requirements.txt
# Основні бібліотеки
pefile>=2023.2.7
yara-python>=4.5.0
psutil>=5.9.8
python-magic-bin>=0.4.14
pillow>=11.0.0
requests>=2.31.0
oletools>=0.60.2
rarfile>=4.2
py7zr>=0.20.5
scikit-learn>=1.4.2
joblib>=1.3.2
pypdf2>=3.0.1
lief>=0.13.2
capstone>=5.0.1
